requirements.txt
asgiref==3.8.1
Django==5.0.4
django-restframework==0.0.1
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
django-filter~=24.2
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
